requirements.txt
fastapi
uvicorn
pydantic
supabase
openai
python-dotenv
streamlit
requests
pandas
plotly
